UF-5930: Acted on PR comments
- Refactored specification to reduce cognitive complexity
- Changed service to throw client fault when deleting a non existing
asset
- Removed unnecessary circuit breaker annotation

+++ src/main/java/se/sundsvall/partyassets/service/AssetService.java
@@ -58,6 +58,13 @@ public class AssetService {
 	}
 
 	public void deleteAsset(final String id) {
+		if (!repository.existsById(id)) {
+			throw Problem.builder()
+				.withStatus(NOT_FOUND)
+				.withTitle("Asset not found")
+				.withDetail("Asset with id %s not found".formatted(id))
+				.build();
+		}
 		repository.deleteById(id);
 	}
